requirements.txt
cdsapi>=0.6.1
dask>=2024.1.0
matplotlib>=3.8.0
netcdf4>=1.6.5
numpy>=1.26.0
pandas>=2.1.0
pyyaml>=6.0.1
scikit-learn>=1.4.0
torch>=2.2.0
tqdm>=4.66.0
xarray>=2024.1.0
zarr>=2.17.0
